+++ src/bitboardhelper.cpp
@@ -38,7 +38,7 @@ uint8_t BBHelper::bit_to_square(uint8_t bit) {
     return GRID_NUM_SQUARES - bit - 1;
 }
 
-void BBHelper::set_bit_by_ref(uint64_t& b, int bit) {
+void BBHelper::set_bit_by_ref(uint64_t& b, uint8_t bit) {
     b |= (1ULL << bit);
 }
